feat: ✨ add support for listing model information in CLI
